SecurityPkg/DxeImageVerificationLib: Remove hard-coded SHA256 to verify unsigned PE/COFF Image
This implementation provides a mechanism to the user to select a strong
hashing algorithm to verify the unsigned PE/COFF image instead of hard-coded
SHA256 algorithm. The appropriate hashing algorithm is provided by the PCD,
PcdSystemHashPolicy.

The user must make sure that the hashing algorithm used to verify a specific
unsigned PE/COFF image that is specified by PcdSystemHashPolicy matches the
one used to hash that image while enrolling the hash into DB/DBX.

Signed-off-by: Sukerkar, Amol N <amol.n.sukerkar@intel.com>